Add support for plain TCP loadbalancing
Please note that to use this you have to use Nginx 1.9 or
a newer one. This adds an nginx_official_repo_mainline variable
that allows using the mainline repo.

To use the stream feature you have to have both nginx_official_repo
and nginx_official_repo_mainline set to true.