Removed keepalive_timeout and types_hash_max_size
keepalive_timeout and types_hash_max_size are performance related settings and it should be left to the admin to decide what their values should be if they are not happy with nginx's built-in defaults. Defining these in `vars/main.yml` causes more pain because it makes it hard to override (the only way is by passing in a new version of `nginx_http_default_params` as a role var or on the command line, which is a strange place to do it given these are supposed to be generic defaults),

Example: I prefer `keepalive_timeout 30 30` for all of my hosts but currently have to define this in every `nginx_sites` entry because I can't set it once via an include using `nginx_configs` (nginx will complain of a duplicate setting and refuse to start).