Newer os compatiblity with os dependent vars (#245)
Added a functionality to the task in main.yml that includes the os
dependent vars whereby you can speficie the os version of the vars that
you want. The corresponding file(s) for that os version must be created in the vars folder. The compatibility with Debian12 was also created in the vars folder.

Co-authored-by: Alex Reid <alex.reid@stonehagefleming.com>